Round amounts in brev to nearest hele krone

+++ domain/src/main/kotlin/no/nav/su/se/bakover/domain/brev/beregning/Månedsfradrag.kt
@@ -4,6 +4,6 @@ import no.nav.su.se.bakover.domain.beregning.fradrag.UtenlandskInntekt
 
 data class Månedsfradrag(
     val type: String,
-    val beløp: Double,
+    val beløp: Int,
     val utenlandskInntekt: UtenlandskInntekt?
 )
